backend/requirements.txt
# Core Flask and web framework
Flask==3.1.0
flask-cors==6.0.0
Werkzeug==3.1.3

# Database
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1

# Authentication
PyJWT==2.10.1
cryptography==44.0.0

# Data processing and analysis
pandas==2.3.0
numpy==2.2.6

# Machine learning
scikit-learn==1.7.0
scipy==1.15.3
joblib==1.5.1

# Utilities
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2
six==1.17.0

# Flask utilities
Jinja2==3.1.4
itsdangerous==2.2.0
click==8.1.8
blinker==1.9.0

# Threading
threadpoolctl==3.6.0

# Development and testing (optional)
# pytest==8.0.0
# pytest-flask==1.3.0
# black==24.0.0
# flake8==7.0.0

# Production server (optional)
# gunicorn==21.2.0
# waitress==3.0.0